moved `data.py` in subpackage `symmetry_function`.
